pyproject.toml
[project]
name = "long-emotion-mc-es"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "flashinfer-cubin>=0.4.1",
    "flashinfer-jit-cache>=0.4.1",
    "flashinfer-python>=0.3.1.post1",
    "vllm>=0.11.0",
    "numpy>=1.26",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.7.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[tool.uv.pip]
torch-backend = "auto"

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[[tool.uv.index]]
name = "flashinfer-cu128"
url = "https://flashinfer.ai/whl/cu128"

[tool.uv.sources]
flashinfer-jit-cache = { index = "flashinfer-cu128" }